+++ internal/dba/server.go
@@ -36,7 +36,7 @@ func (s *DbaServer) Handler(work *base.Work) {
 		logger.Error("Invalid data from work struct.")
 		return
 	}
-	switch byte(agreement.Cmd) {
+	switch agreement.Cmd {
 	case define.SystemCommand:
 		s.handleSystemCommand(work, agreement)
 	case define.NormalCommand:
@@ -54,13 +54,11 @@ func (s *DbaServer) Run() {
 }
 
 func (s *DbaServer) handleSystemCommand(work *base.Work, agreement *agrt.Agreement) {
-	switch uint16(agreement.Service) {
+	switch agreement.Service {
 	// 回應心跳包
 	case define.Heartbeat:
 		fmt.Printf("Heart beat! Now: %+v\n", time.Now())
 		work.Body.Clear()
-		// work.Body.AddByte(define.SystemCommand)
-		// work.Body.AddUInt16(define.Heartbeat)
 		bs, _ := agreement.Marshal()
 		work.Body.AddByteArray(bs)
 		work.SendTransData()
@@ -71,7 +69,7 @@ func (s *DbaServer) handleSystemCommand(work *base.Work, agreement *agrt.Agreeme
 }
 
 func (s *DbaServer) handleNormalCommand(work *base.Work, agreement *agrt.Agreement) {
-	switch uint16(agreement.Service) {
+	switch agreement.Service {
 	case define.GetUserData:
 		logger.Debug("GetUserData")
 		work.Body.Clear()
@@ -85,13 +83,12 @@ func (s *DbaServer) handleNormalCommand(work *base.Work, agreement *agrt.Agreeme
 		defer s.tables[TidAccount].PutSelector(selector)
 		results, err := selector.Query(func() any { return &pbgo.Account{} })
 		if err != nil {
-			// work.Body.AddUInt16(1)
 			agreement.ReturnCode = 1
 			agreement.Msg = "Failed to select data."
 			logger.Error("Select err: %+v", err)
 			return
 		}
-		// accounts := &pbgo.AccountArray{}
+
 		var account *pbgo.Account
 		for _, result := range results {
 			account = result.(*pbgo.Account)
@@ -99,34 +96,15 @@ func (s *DbaServer) handleNormalCommand(work *base.Work, agreement *agrt.Agreeme
 			logger.Debug("account: %+v", account)
 			agreement.Accounts = append(agreement.Accounts, account)
 		}
-		// bs, err := proto.Marshal(accounts)
-		// if err != nil {
-		// 	logger.Error("Select err: %+v", err)
-		// 	work.Body.AddUInt16(2)
-		// 	return
-		// }
 
 		agreement.ReturnCode = 0
-		// work.Body.AddUInt16(0)
-		// work.Body.AddByteArray(bs)
 	}
 }
 
 func (s *DbaServer) handleCommission(work *base.Work, agreement *agrt.Agreement) {
-	// commission := work.Body.PopUInt16()
-	// cid := work.Body.PopInt32()
-	// logger.Info("commission: %d, cid: %d", commission, cid)
-
-	switch uint16(agreement.Service) {
+	switch agreement.Service {
 	case define.Register:
-		// 建立使用者資料
-		// bs := work.Body.PopByteArray()
-		// account := &pbgo.Account{}
-		// err := proto.Unmarshal(bs, account)
 		work.Body.Clear()
-		// work.Body.AddByte(define.CommissionCommand)
-		// work.Body.AddUInt16(define.Register)
-		// work.Body.AddInt32(cid)
 		defer func() {
 			bs, _ := agreement.Marshal()
 			work.Body.AddByteArray(bs)
@@ -157,19 +135,10 @@ func (s *DbaServer) handleCommission(work *base.Work, agreement *agrt.Agreement)
 		logger.Info("result: %s", result)
 		// returnCode
 		agreement.ReturnCode = 0
-		// work.Body.AddUInt16(0)
 		account.Index = int32(result.LastInsertId)
 
 	case define.SetUserData:
-		// 建立使用者資料
-		// bs := work.Body.PopByteArray()
-		// account := &pbgo.Account{}
-		// err := proto.Unmarshal(bs, account)
-		// logger.Info("account: %+v", account)
 		work.Body.Clear()
-		// work.Body.AddByte(define.CommissionCommand)
-		// work.Body.AddUInt16(define.SetUserData)
-		// work.Body.AddInt32(cid)
 		defer func() {
 			bs, _ := agreement.Marshal()
 			work.Body.AddByteArray(bs)
@@ -193,8 +162,6 @@ func (s *DbaServer) handleCommission(work *base.Work, agreement *agrt.Agreement)
 
 		// returnCode
 		agreement.ReturnCode = 0
-		// work.Body.AddUInt16(0)
-		// work.Body.AddByteArray(bs)
 
 	default:
 		fmt.Printf("Unsupport commission: %d\n", agreement.Service)
